MAINT: move ConstantFunction
